Remove redundant data C files from source control
blank_jpeg.c and index_html.c are machine-generated files that we derive from other files in source. We don't need to store the machine-generated versions in source control, as we can generate them on-demand.

This change deletes blank_jpeg.c and index_html.c, adds them to .gitignore, and updates the Makefile to build them as needed.