Clarify that dual_final_frames parameter applies to most browsers
The current documentation says that this option is specifically for Safari and WebKit-based browsers, but I see this bug on Windows versions of Chrome, Firefox, and Edge.

This change updates the documentation for the dual_final_frames URL parameter to clarify that the workaround applies to many browsers beyond Safari.